Add bluetooth_le permission, fix kernel bluetooh osi fail
At the nougat age, there is an issue for some kernel which didn't
support CLOCK_BOOTTIME_ALARM or when RTC alarm is presenst. It would
cause the kernel bluetooth osi fail and deadlock at the Manufacturer's
logo.
There is a message in the last_kmsg which tells you the problem. I used
some unusal method to fix it earlier but turn out there is a flag in the
new nougat source to handle that.
Set KERNEL_MISSING_CLOCK_BOOTTIME_ALARM to TRUE in your
BOARD_BLUETOOTH_BDROID_BUILDCFG_INCLUDE_DIR file will fix it.